requirements.txt
Flask>=1,<2
werkzeug>=1,<2
flask-session~=0.3.2
requests>=2,<3
msal>=0.6.1,<2
yapf>=0.28.0
opencensus>=0.7.7
opencensus-context>=0.1.1
opencensus-ext-azure>=1.0.2
opencensus-ext-flask>=0.7.3
azure.identity>=1.3.0
azure-keyvault-secrets>=4.1.0
flasgger>=0.9.3
PyYAML>=5.1.2
pylint>=2.3.1
watchdog>=0.9.0